+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.1.3] - 2021-03-19
+### Changed
+- Fixes bug in NextJS wrapper which called resolve even though there was an error
+
 ## [4.1.2] - 2021-03-05
 ### Changed
 - Fix Domain normalisation with "/.netlify/functions/api"